fix: separate domain objects to diff files
