+++ internal/domain/rank.go
@@ -1,0 +1,12 @@
+package domain
+
+import "github.com/google/uuid"
+
+type Rank struct {
+	ID   uuid.UUID `json:"id" gorm:"type:uuid;default:uuid_generate_v4()"`
+	Name string    `json:"name" gorm:"unique;not null"`
+}
+
+func (Rank) TableName() string {
+	return "ranks"
+}
